APMSPII-376 Add per-environment hooks _scoped_pipeline_vars
The azure pipline template expansion seems to have a notion of a
"current directory" and "current repository". In the case that the
stepLists are templated, the tempalate expansion now happens in the
common repository so unless the user specifies a repo, we assume it
happens in '@self' in the azure/ directory. This means the behaviour
of paths are the same whether or not the hook is defined at a global
or per-environment scope.